Use MY_GITHUB_TOKEN instead of GITHUB_TOKEN
Since the project that we're adding issues to isn't in the
eclipse-pass/main repo, the default GITHUB_TOKEN doesn't have
authorization to add issues to it. Instead, I've generated a token
for my own user and had that token added to the repo as a secret
named MY_GITHUB_TOKEN. The srggrs/assign-one-project-github-action
workflow automatically looks for the MY_GITHUB_TOKEN when the specified
project is outside the repo.